Moving tests of computed_by feature into computed_by_test.go
